🧪 FD8ZPX tests: split guard commit wrapper suite
