+++ .agentplane/tasks/202605010645-FD8ZPX/pr/github-body.md
@@ -1,0 +1,33 @@
+## Summary
+
+AP-11: Split guard commit wrapper tests
+
+Split guard commit-wrapper CLI scenarios by env, policy, close, and refresh behavior with shared testkit fixtures.
+
+## Scope
+
+- In scope: Split guard commit-wrapper CLI scenarios by env, policy, close, and refresh behavior with shared testkit fixtures.
+- Out of scope: unrelated refactors not required for "AP-11: Split guard commit wrapper tests".
+
+## Verification
+
+- State: ok
+- Note: Verified: guard commit-wrapper tests split into env, policy, refresh, and close suites; routing, inventory, and significant coverage use the split files; oversized baseline ratcheted.
+- Full verification checklist lives in local review.md.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-01T10:33:13.996Z
+- Branch: task/202605010645-FD8ZPX/guard-commit-wrapper-test-split
+- Head: fc85c7d5cee5
+
+```text
+No changes detected.
+```
+
+</details>
