+++ .agentplane/tasks/202605010645-FD8ZPX/pr/github-title.txt
@@ -1,0 +1,1 @@
+AP-11: Split guard commit wrapper tests (FD8ZPX)
