--- packages/agentplane/src/cli/run-cli.core.guard.commit-wrapper.test.ts
@@ -1,1217 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import { execFile } from "node:child_process";
-import { readFileSync } from "node:fs";
-import {
-  chmod,
-  mkdir,
-  mkdtemp,
-  readdir,
-  readFile,
-  realpath,
-  rm,
-  writeFile,
-} from "node:fs/promises";
-import os from "node:os";
-import path from "node:path";
-import { promisify } from "node:util";
-import { describe, expect, it, vi } from "vitest";
-import { defaultConfig, extractTaskSuffix, type ResolvedProject } from "./core-imports.js";
-import { readTask, renderTaskReadme } from "@agentplaneorg/core/tasks";
-
-import { runCli } from "./run-cli.js";
-import {
-  filterAgentsByWorkflow,
-  loadAgentTemplates,
-  loadAgentsTemplate,
-} from "../agents/agents-template.js";
-import type * as taskBackend from "../backends/task-backend.js";
-import {
-  captureStdIO,
-  cleanGitEnv,
-  commitAll,
-  configureGitUser,
-  createUpgradeBundle,
-  getAgentplaneHome,
-  gitBranchExists,
-  installRunCliIntegrationHarness,
-  runCliSilent,
-  mkGitRepoRoot,
-  mkGitRepoRootWithBranch,
-  mkTempDir,
-  pathExists,
-  stageGitignoreIfPresent,
-  stubTaskBackend,
-  writeConfig,
-  writeDefaultConfig,
-} from "@agentplane/testkit";
-import { resolveUpdateCheckCachePath } from "./update-check.js";
-import * as prompts from "./prompts.js";
-
-installRunCliIntegrationHarness();
-
-const COMMIT_WRAPPER_AUTO_STAGE_TIMEOUT_MS = 120_000;
-const COMMIT_WRAPPER_CLOSE_CHECK_ONLY_TIMEOUT_MS = 120_000;
-const COMMIT_WRAPPER_SUITE_TIMEOUT_MS = 120_000;
-
-describe("runCli", { timeout: COMMIT_WRAPPER_SUITE_TIMEOUT_MS }, () => {
-  it("commit wrapper creates a commit with allowlist", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    await writeFile(path.join(root, "file.txt"), "x", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "file.txt"], { cwd: root });
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        "202601010101-ABCDEF",
-        "-m",
-        "✨ ABCDEF commit: wrapper command",
-        "--allow",
-        "file.txt",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      expect(io.stdout).toContain("✅ committed");
-    } finally {
-      io.restore();
-    }
-
-    const { stdout } = await execFileAsync("git", ["log", "-1", "--pretty=%s"], { cwd: root });
-    expect(stdout.trim()).toBe("✨ ABCDEF commit: wrapper command");
-  });
-
-  it("commit wrapper auto-stages allowlist-scoped changes when the index is empty", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    await mkdir(path.join(root, "src"), { recursive: true });
-    await writeFile(path.join(root, "src", "app.ts"), "export const x = 1;\n", "utf8");
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        "202601010101-ABCDEF",
-        "-m",
-        "✨ ABCDEF commit: auto stage allowlist",
-        "--allow",
-        "src",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      expect(io.stdout).toContain("commit auto-staged 1 path(s) from allowlist");
-      expect(io.stdout).toContain("staged=src/app.ts");
-    } finally {
-      io.restore();
-    }
-
-    const execFileAsync = promisify(execFile);
-    const { stdout } = await execFileAsync("git", ["log", "-1", "--pretty=%s"], { cwd: root });
-    expect(stdout.trim()).toBe("✨ ABCDEF commit: auto stage allowlist");
-  });
-
-  it(
-    "commit wrapper auto-stages active task artifacts with --allow-tasks even without explicit --allow",
-    async () => {
-      const root = await mkGitRepoRoot();
-      await writeDefaultConfig(root);
-      await configureGitUser(root);
-      const taskId = "202601010101-ABCDEF";
-      const readme = renderTaskReadme(
-        {
-          id: taskId,
-          title: "Auto-stage active task artifact commit",
-          result_summary: "",
-          description: "desc",
-          status: "DOING",
-          priority: "med",
-          owner: "CODER",
-          depends_on: [],
-          tags: ["cli", "code"],
-          verify: [],
-          verification: null,
-          commit: null,
-          doc_version: 3,
-          doc_updated_at: "2026-03-12T00:00:00.000Z",
-          doc_updated_by: "CODER",
-        },
-        "## Summary\n\nUpdated\n",
-      );
-      await mkdir(path.join(root, ".agentplane", "tasks", taskId), { recursive: true });
-      await writeFile(path.join(root, ".agentplane", "tasks", taskId, "README.md"), readme, "utf8");
-
-      const io = captureStdIO();
-      try {
-        const code = await runCli([
-          "commit",
-          taskId,
-          "-m",
-          "✨ ABCDEF commit: auto stage active task artifact",
-          "--allow-tasks",
-          "--root",
-          root,
-        ]);
-        expect(code).toBe(0);
-        expect(io.stdout).toContain("commit auto-staged 1 path(s) from allowlist");
-        expect(io.stdout).toContain(`staged=.agentplane/tasks/${taskId}/README.md`);
-      } finally {
-        io.restore();
-      }
-
-      const execFileAsync = promisify(execFile);
-      const { stdout } = await execFileAsync("git", ["log", "-1", "--pretty=%s"], { cwd: root });
-      expect(stdout.trim()).toBe("✨ ABCDEF commit: auto stage active task artifact");
-    },
-    COMMIT_WRAPPER_AUTO_STAGE_TIMEOUT_MS,
-  );
-
-  it(
-    "commit wrapper auto-stages CI-only changes with --allow-ci even without explicit --allow",
-    async () => {
-      const root = await mkGitRepoRoot();
-      await writeDefaultConfig(root);
-      await configureGitUser(root);
-      await mkdir(path.join(root, ".github", "workflows"), { recursive: true });
-      await writeFile(
-        path.join(root, ".github", "workflows", "publish.yml"),
-        "name: publish\n",
-        "utf8",
-      );
-
-      const io = captureStdIO();
-      try {
-        const code = await runCli([
-          "commit",
-          "202601010101-ABCDEF",
-          "-m",
-          "✨ ABCDEF commit: auto stage ci artifact",
-          "--allow-ci",
-          "--root",
-          root,
-        ]);
-        expect(code).toBe(0);
-        expect(io.stdout).toContain("commit auto-staged 1 path(s) from allowlist");
-        expect(io.stdout).toContain("staged=.github/workflows/publish.yml");
-      } finally {
-        io.restore();
-      }
-
-      const execFileAsync = promisify(execFile);
-      const { stdout } = await execFileAsync("git", ["log", "-1", "--pretty=%s"], { cwd: root });
-      expect(stdout.trim()).toBe("✨ ABCDEF commit: auto stage ci artifact");
-    },
-    COMMIT_WRAPPER_AUTO_STAGE_TIMEOUT_MS,
-  );
-
-  it("commit wrapper blocks AGENTS.md without allow-policy", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    await writeFile(path.join(root, "AGENTS.md"), "# policy\n", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "AGENTS.md"], { cwd: root });
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        "202601010101-ABCDEF",
-        "-m",
-        "✨ ABCDEF commit: protected policy",
-        "--allow",
-        "AGENTS.md",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(5);
-      expect(io.stderr).toContain("AGENTS.md");
-      expect(io.stderr).toContain("--allow-policy");
-    } finally {
-      io.restore();
-    }
-  });
-
-  it("commit wrapper allows AGENTS.md with allow-policy", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    await writeFile(path.join(root, "AGENTS.md"), "# policy\n", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "AGENTS.md"], { cwd: root });
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        "202601010101-ABCDEF",
-        "-m",
-        "✨ ABCDEF commit: protected policy",
-        "--allow",
-        "AGENTS.md",
-        "--allow-policy",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      expect(io.stdout).toContain("✅ committed");
-    } finally {
-      io.restore();
-    }
-
-    const { stdout } = await execFileAsync("git", ["log", "-1", "--pretty=%s"], { cwd: root });
-    expect(stdout.trim()).toBe("✨ ABCDEF commit: protected policy");
-  });
-
-  it("commit wrapper normalizes ./ prefixes in allowlist", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    await writeFile(path.join(root, "file.txt"), "x", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "file.txt"], { cwd: root });
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        "202601010101-ABCDEF",
-        "-m",
-        "✨ ABCDEF commit: allow file.txt",
-        "--allow",
-        "./file.txt",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      expect(io.stdout).toContain("✅ committed");
-    } finally {
-      io.restore();
-    }
-  });
-
-  it("commit wrapper rejects repo-wide allowlist prefix", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    await writeFile(path.join(root, "file.txt"), "x", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "file.txt"], { cwd: root });
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        "202601010101-ABCDEF",
-        "-m",
-        "✨ ABCDEF commit: reject repo wide",
-        "--allow",
-        ".",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(2);
-      expect(io.stderr).toContain("cannot be repo-wide");
-    } finally {
-      io.restore();
-    }
-  });
-
-  it("commit wrapper supports --quiet", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    await writeFile(path.join(root, "file.txt"), "x", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "file.txt"], { cwd: root });
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        "202601010101-ABCDEF",
-        "-m",
-        "✨ ABCDEF commit: quiet mode",
-        "--allow",
-        "file.txt",
-        "--quiet",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      expect(io.stdout.trim()).toBe("");
-    } finally {
-      io.restore();
-    }
-  });
-
-  it("commit wrapper rejects auto-allow", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    await mkdir(path.join(root, "src"), { recursive: true });
-    await writeFile(path.join(root, "src", "app.ts"), "x", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "src/app.ts"], { cwd: root });
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        "202601010101-ABCDEF",
-        "-m",
-        "✨ ABCDEF commit: auto allow",
-        "--auto-allow",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(2);
-      expect(io.stderr).toContain("--auto-allow is disabled");
-    } finally {
-      io.restore();
-    }
-  });
-
-  it("commit wrapper compresses noisy pre-commit output into summary", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    await writeFile(path.join(root, "file.txt"), "x", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "file.txt"], { cwd: root });
-
-    const hookLines = Array.from({ length: 40 }, (_, i) =>
-      i === 19
-        ? "\u001B[33mCode style issues found in the above file. Run Prettier with --write to fix.\u001B[0m"
-        : `HOOK_LINE_${String(i + 1).padStart(2, "0")}`,
-    );
-    const preCommit = [
-      "#!/bin/sh",
-      ...hookLines.map((line) => `echo "${line}" 1>&2`),
-      "exit 1",
-      "",
-    ].join("\n");
-    const hookPath = path.join(root, ".git", "hooks", "pre-commit");
-    await writeFile(hookPath, preCommit, "utf8");
-    await chmod(hookPath, 0o755);
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        "202601010101-ABCDEF",
-        "-m",
-        "✨ ABCDEF commit: hook failure summary",
-        "--allow",
-        "file.txt",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(5);
-      expect(io.stderr).toContain("error [E_GIT]");
-      expect(io.stderr).toContain("git commit failed");
-      expect(io.stderr).toContain("state: git rejected the requested task-scoped commit");
-      expect(io.stderr).toContain(
-        "likely_cause: a formatting check in the pre-commit path rejected the staged task-scoped commit",
-      );
-      expect(io.stderr).toContain(
-        "next_action: bun run format (apply formatter fixes before retrying the commit flow)",
-      );
-      expect(io.stderr).toContain("reason_code: git_pre_commit_format");
-      expect(io.stderr).toContain("output_summary:");
-      expect(io.stderr).toContain("HOOK_LINE_01");
-      expect(io.stderr).toContain(
-        "Code style issues found in the above file. Run Prettier with --write to fix.",
-      );
-      expect(io.stderr).toContain("HOOK_LINE_40");
-      expect(io.stderr).toContain("lines omitted");
-      expect(io.stderr).not.toContain("HOOK_LINE_20");
-      expect(io.stderr).not.toContain("\u001B[33m");
-      expect(io.stderr).not.toContain("generated close commit");
-    } finally {
-      io.restore();
-    }
-  });
-
-  it("commit wrapper supports --allow-base and --require-clean", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    await writeFile(path.join(root, "base.txt"), "x", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "base.txt"], { cwd: root });
-    // Unstaged tracked changes should block --require-clean.
-    await writeFile(path.join(root, "base.txt"), "y", "utf8");
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        "202601010101-ABCDEF",
-        "-m",
-        "✨ ABCDEF commit: allow base",
-        "--allow",
-        "base.txt",
-        "--allow-base",
-        "--require-clean",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(5);
-    } finally {
-      io.restore();
-    }
-  });
-
-  it("commit wrapper supports --allow-tasks flag", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    await writeFile(path.join(root, "note.txt"), "x", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "note.txt"], { cwd: root });
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        "202601010101-ABCDEF",
-        "-m",
-        "✨ ABCDEF commit: allow tasks",
-        "--allow",
-        "note.txt",
-        "--allow-tasks",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-    } finally {
-      io.restore();
-    }
-  });
-
-  it("commit wrapper lets --allow-tasks cover the active task README without a duplicate explicit prefix", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    const taskId = "202601010101-ABCDEF";
-    const readme = renderTaskReadme(
-      {
-        id: taskId,
-        title: "Allow active task artifacts in task-scoped commits",
-        result_summary: "",
-        description: "desc",
-        status: "DOING",
-        priority: "med",
-        owner: "CODER",
-        depends_on: [],
-        tags: ["cli", "code"],
-        verify: [],
-        verification: null,
-        commit: null,
-        doc_version: 3,
-        doc_updated_at: "2026-03-12T00:00:00.000Z",
-        doc_updated_by: "CODER",
-      },
-      "## Summary\n\nUpdated\n",
-    );
-    await mkdir(path.join(root, "src"), { recursive: true });
-    await mkdir(path.join(root, ".agentplane", "tasks", taskId), { recursive: true });
-    await writeFile(path.join(root, "src", "app.ts"), "export const x = 1;\n", "utf8");
-    await writeFile(path.join(root, ".agentplane", "tasks", taskId, "README.md"), readme, "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "src/app.ts", `.agentplane/tasks/${taskId}/README.md`], {
-      cwd: root,
-    });
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        taskId,
-        "-m",
-        "✨ ABCDEF commit: allow active task artifacts",
-        "--allow",
-        "src",
-        "--allow-tasks",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      expect(io.stdout).toContain("committed");
-    } finally {
-      io.restore();
-    }
-
-    const showRes = await execFileAsync("git", ["show", "--name-only", "--format="], {
-      cwd: root,
-    });
-    const changed = showRes.stdout
-      .split("\n")
-      .map((s) => s.trim())
-      .filter(Boolean);
-    expect(changed).toEqual([`.agentplane/tasks/${taskId}/README.md`, "src/app.ts"]);
-  });
-
-  it("commit wrapper stages active task artifacts with --allow-tasks when the index already has implementation files", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    const taskId = "202601010101-ABCDEF";
-    const readme = renderTaskReadme(
-      {
-        id: taskId,
-        title: "Stage active task artifacts with populated index",
-        result_summary: "",
-        description: "desc",
-        status: "DOING",
-        priority: "med",
-        owner: "CODER",
-        depends_on: [],
-        tags: ["cli", "code"],
-        verify: [],
-        verification: null,
-        commit: null,
-        doc_version: 3,
-        doc_updated_at: "2026-03-12T00:00:00.000Z",
-        doc_updated_by: "CODER",
-      },
-      "## Summary\n\nUpdated\n",
-    );
-    await mkdir(path.join(root, "src"), { recursive: true });
-    await mkdir(path.join(root, ".agentplane", "tasks", taskId), { recursive: true });
-    await writeFile(path.join(root, "src", "app.ts"), "export const x = 1;\n", "utf8");
-    await writeFile(path.join(root, ".agentplane", "tasks", taskId, "README.md"), readme, "utf8");
-    await writeFile(
-      path.join(root, ".agentplane", "tasks", taskId, "evidence.txt"),
-      "active task evidence\n",
-      "utf8",
-    );
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "src/app.ts"], { cwd: root });
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        taskId,
-        "-m",
-        "✨ ABCDEF commit: stage active task artifacts with populated index",
-        "--allow",
-        "src",
-        "--allow-tasks",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      expect(io.stdout).toContain(
-        "commit auto-staged 2 active task artifact path(s) from --allow-tasks",
-      );
-      expect(io.stdout).toContain(
-        `staged=.agentplane/tasks/${taskId}/evidence.txt, .agentplane/tasks/${taskId}/README.md`,
-      );
-    } finally {
-      io.restore();
-    }
-
-    const showRes = await execFileAsync("git", ["show", "--name-only", "--format="], {
-      cwd: root,
-    });
-    const changed = showRes.stdout
-      .split("\n")
-      .map((s) => s.trim())
-      .filter(Boolean);
-    expect(changed).toEqual([
-      `.agentplane/tasks/${taskId}/README.md`,
-      `.agentplane/tasks/${taskId}/evidence.txt`,
-      "src/app.ts",
-    ]);
-  });
-
-  it("commit wrapper lets --allow-tasks cover a non-README active task artifact without a duplicate explicit prefix", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    const taskId = "202601010101-ABCDEF";
-    const readme = renderTaskReadme(
-      {
-        id: taskId,
-        title: "Allow non-README active task artifacts in task-scoped commits",
-        result_summary: "",
-        description: "desc",
-        status: "DOING",
-        priority: "med",
-        owner: "CODER",
-        depends_on: [],
-        tags: ["cli", "code"],
-        verify: [],
-        verification: null,
-        commit: null,
-        doc_version: 3,
-        doc_updated_at: "2026-03-12T00:00:00.000Z",
-        doc_updated_by: "CODER",
-      },
-      "## Summary\n\nUpdated\n",
-    );
-    await mkdir(path.join(root, "src"), { recursive: true });
-    await mkdir(path.join(root, ".agentplane", "tasks", taskId), { recursive: true });
-    await writeFile(path.join(root, "src", "app.ts"), "export const x = 1;\n", "utf8");
-    await writeFile(path.join(root, ".agentplane", "tasks", taskId, "README.md"), readme, "utf8");
-    await writeFile(
-      path.join(root, ".agentplane", "tasks", taskId, "evidence.txt"),
-      "active task evidence\n",
-      "utf8",
-    );
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "src/app.ts", `.agentplane/tasks/${taskId}/evidence.txt`], {
-      cwd: root,
-    });
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        taskId,
-        "-m",
-        "✨ ABCDEF commit: allow active task evidence artifact",
-        "--allow",
-        "src",
-        "--allow-tasks",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      expect(io.stdout).toContain("committed");
-    } finally {
-      io.restore();
-    }
-
-    const headRes = await execFileAsync("git", ["show", "--name-only", "--format="], {
-      cwd: root,
-    });
-    const headChanged = headRes.stdout
-      .split("\n")
-      .map((s) => s.trim())
-      .filter(Boolean);
-    expect(headChanged).toEqual([
-      `.agentplane/tasks/${taskId}/README.md`,
-      `.agentplane/tasks/${taskId}/evidence.txt`,
-      "src/app.ts",
-    ]);
-  });
-
-  it("commit wrapper surfaces invalid task README frontmatter as the concrete reconcile failure", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    const taskId = "202601010101-ABCDEF";
-    await mkdir(path.join(root, "src"), { recursive: true });
-    await mkdir(path.join(root, ".agentplane", "tasks", taskId), { recursive: true });
-    await writeFile(path.join(root, "src", "app.ts"), "export const x = 1;\n", "utf8");
-    await writeFile(
-      path.join(root, ".agentplane", "tasks", taskId, "README.md"),
-      "# Task\n",
-      "utf8",
-    );
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "src/app.ts"], { cwd: root });
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        taskId,
-        "-m",
-        "✨ ABCDEF commit: blocked by invalid task readme",
-        "--allow",
-        "src",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(3);
-      expect(io.stderr).toContain(
-        "reconcile check failed: task README for 202601010101-ABCDEF has invalid frontmatter and could not be parsed",
-      );
-      expect(io.stderr).toContain("state: mutation preflight cannot reconcile task artifacts");
-      expect(io.stderr).toContain(
-        "likely_cause: task README for 202601010101-ABCDEF has invalid frontmatter, so reconcile skipped it before the mutating command could run",
-      );
-      expect(io.stderr).toContain(
-        "next_action: agentplane task list --strict-read (surface the malformed or unreadable task README before retrying mutating commands)",
-      );
-      expect(io.stderr).toContain("reason_code: reconcile_task_scan_incomplete [reconcile]");
-    } finally {
-      io.restore();
-    }
-  });
-
-  it(
-    "commit wrapper supports --close and stages only the task README",
-    async () => {
-      const root = await mkGitRepoRoot();
-      await writeDefaultConfig(root);
-      await configureGitUser(root);
-
-      // Seed the repo so the implementation commit has a parent.
-      await writeFile(path.join(root, "seed.txt"), "seed\n", "utf8");
-      const execFileAsync = promisify(execFile);
-      await execFileAsync("git", ["add", "seed.txt"], { cwd: root });
-      await execFileAsync("git", ["commit", "-m", "seed"], { cwd: root });
-
-      // Implementation commit recorded into the task metadata.
-      await mkdir(path.join(root, "src"), { recursive: true });
-      await writeFile(
-        path.join(root, "src", "app.ts"),
-        ["x", "x", "x", "x", "x"].join("\n"),
-        "utf8",
-      );
-      await execFileAsync("git", ["add", "src/app.ts"], { cwd: root });
-      await execFileAsync("git", ["commit", "-m", "impl"], { cwd: root });
-      const implRes = await execFileAsync("git", ["rev-parse", "HEAD"], { cwd: root });
-      const implHash = implRes.stdout.trim();
-
-      const taskId = "202602081506-R18Y1Q";
-      const taskDir = path.join(root, ".agentplane", "tasks", taskId);
-      await mkdir(taskDir, { recursive: true });
-      const readme = renderTaskReadme(
-        {
-          id: taskId,
-          title: "Close commit: add --close mode to agentplane commit",
-          result_summary: "generate deterministic close commits",
-          description: "desc",
-          status: "DONE",
-          priority: "high",
-          owner: "ORCHESTRATOR",
-          depends_on: [],
-          tags: ["cli", "code", "git"],
-          verify: ["bun run test:full"],
-          verification: {
-            state: "ok",
-            updated_at: "2026-02-08T00:00:00.000Z",
-            updated_by: "TESTER",
-            note: "Verified: bun run test:full; manual: agentplane commit --close",
-          },
-          commit: { hash: implHash, message: "✨ R18Y1Q guard: add close mode" },
-          doc_version: 3,
-          doc_updated_at: "2026-02-08T00:00:00.000Z",
-          doc_updated_by: "TESTER",
-        },
-        "## Summary\n\nTest task\n",
-      );
-      await writeFile(path.join(taskDir, "README.md"), readme, "utf8");
-
-      const io = captureStdIO();
-      try {
-        const code = await runCli(["commit", taskId, "--close", "--root", root]);
-        expect(code).toBe(0);
-        expect(io.stdout).toContain("committed");
-      } finally {
-        io.restore();
-      }
-
-      const subjectRes = await execFileAsync("git", ["log", "-1", "--pretty=%s"], { cwd: root });
-      const bodyRes = await execFileAsync("git", ["log", "-1", "--pretty=%b"], { cwd: root });
-      const subject = subjectRes.stdout.trim();
-      const body = bodyRes.stdout.trim();
-      expect(subject).toContain("✅ R18Y1Q close:");
-      expect(subject).toContain("(202602081506-R18Y1Q)");
-      expect(subject).toContain("[cli,code,git]");
-      expect(body).toContain("Scope: cli, code, git");
-      expect(body).toContain(
-        "Verify: Verified: bun run test:full; manual: agentplane commit --close",
-      );
-      expect(body).toContain("Key files: src/app.ts");
-
-      // Close commit should touch only the task README.
-      const showRes = await execFileAsync("git", ["show", "--name-only", "--format="], {
-        cwd: root,
-      });
-      const changed = showRes.stdout
-        .split("\n")
-        .map((s) => s.trim())
-        .filter(Boolean);
-      expect(changed).toEqual([`.agentplane/tasks/${taskId}/README.md`]);
-    },
-    COMMIT_WRAPPER_CLOSE_CHECK_ONLY_TIMEOUT_MS,
-  );
-
-  it("commit wrapper --close rejects non-empty index", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await configureGitUser(root);
-    await writeFile(path.join(root, "seed.txt"), "seed\n", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "seed.txt"], { cwd: root });
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli(["commit", "202602081506-R18Y1Q", "--close", "--root", root]);
-      expect(code).toBe(5);
-      expect(io.stderr).toContain("close commit requires an empty index");
-      expect(io.stderr).toContain("state: close commit cannot run with a non-empty git index");
-      expect(io.stderr).toContain(
-        "likely_cause: deterministic close commits only stage the active task artifact scope, but other staged files are already in the index",
-      );
-      expect(io.stderr).toContain("next_action: git restore --staged -- .");
-      expect(io.stderr).toContain("--unstage-others");
-    } finally {
-      io.restore();
-    }
-  });
-
-  it(
-    "commit wrapper --close supports --unstage-others",
-    async () => {
-      const root = await mkGitRepoRoot();
-      await writeDefaultConfig(root);
-      await configureGitUser(root);
-      const execFileAsync = promisify(execFile);
-
-      await writeFile(path.join(root, "seed.txt"), "seed\n", "utf8");
-      await execFileAsync("git", ["add", "seed.txt"], { cwd: root });
-      await execFileAsync("git", ["commit", "-m", "seed"], { cwd: root });
-
-      await writeFile(path.join(root, "src.txt"), "impl\n", "utf8");
-      await execFileAsync("git", ["add", "src.txt"], { cwd: root });
-      await execFileAsync("git", ["commit", "-m", "impl"], { cwd: root });
-      const implRes = await execFileAsync("git", ["rev-parse", "HEAD"], { cwd: root });
-      const implHash = implRes.stdout.trim();
-
-      const taskId = "202602081506-R18Y1Q";
-      const taskDir = path.join(root, ".agentplane", "tasks", taskId);
-      await mkdir(taskDir, { recursive: true });
-      const readme = renderTaskReadme(
-        {
-          id: taskId,
-          title: "Close commit with index cleanup",
-          result_summary: "close duplicate bookkeeping",
-          description: "desc",
-          status: "DONE",
-          priority: "high",
-          owner: "ORCHESTRATOR",
-          depends_on: [],
-          tags: ["docs"],
-          verify: ["manual"],
-          verification: {
-            state: "ok",
-            updated_at: "2026-02-08T00:00:00.000Z",
-            updated_by: "TESTER",
-            note: "Verified: manual check",
-          },
-          commit: { hash: implHash, message: "✨ R18Y1Q docs: impl" },
-          doc_version: 3,
-          doc_updated_at: "2026-02-08T00:00:00.000Z",
-          doc_updated_by: "TESTER",
-        },
-        "## Summary\n\nTest task\n",
-      );
-      await writeFile(path.join(taskDir, "README.md"), readme, "utf8");
-
-      await writeFile(path.join(root, "noise.txt"), "noise\n", "utf8");
-      await execFileAsync("git", ["add", "noise.txt"], { cwd: root });
-
-      const io = captureStdIO();
-      try {
-        const code = await runCli([
-          "commit",
-          taskId,
-          "--close",
-          "--unstage-others",
-          "--root",
-          root,
-        ]);
-        expect(code).toBe(0);
-        expect(io.stdout).toContain("committed");
-      } finally {
-        io.restore();
-      }
-
-      const showRes = await execFileAsync("git", ["show", "--name-only", "--format="], {
-        cwd: root,
-      });
-      const changed = showRes.stdout
-        .split("\n")
-        .map((s) => s.trim())
-        .filter(Boolean);
-      expect(changed).toEqual([`.agentplane/tasks/${taskId}/README.md`]);
-    },
-    COMMIT_WRAPPER_CLOSE_CHECK_ONLY_TIMEOUT_MS,
-  );
-
-  it(
-    "commit wrapper --close tolerates dirty README from another active task in direct mode",
-    async () => {
-      const root = await mkGitRepoRoot();
-      await writeDefaultConfig(root);
-      await configureGitUser(root);
-      const execFileAsync = promisify(execFile);
-
-      await writeFile(path.join(root, "seed.txt"), "seed\n", "utf8");
-      await execFileAsync("git", ["add", "seed.txt"], { cwd: root });
-      await execFileAsync("git", ["commit", "-m", "seed"], { cwd: root });
-
-      await writeFile(path.join(root, "src.txt"), "impl\n", "utf8");
-      await execFileAsync("git", ["add", "src.txt"], { cwd: root });
-      await execFileAsync("git", ["commit", "-m", "impl"], { cwd: root });
-      const implRes = await execFileAsync("git", ["rev-parse", "HEAD"], { cwd: root });
-      const implHash = implRes.stdout.trim();
-
-      const closeTaskId = "202602081506-R18Y1Q";
-      const otherTaskId = "202602081506-QA9T2B";
-      const closeTaskDir = path.join(root, ".agentplane", "tasks", closeTaskId);
-      const otherTaskDir = path.join(root, ".agentplane", "tasks", otherTaskId);
-      await mkdir(closeTaskDir, { recursive: true });
-      await mkdir(otherTaskDir, { recursive: true });
-
-      const closeReadme = renderTaskReadme(
-        {
-          id: closeTaskId,
-          title: "Close commit tolerates parallel active task dirt",
-          result_summary: "close one task without swallowing another active task README",
-          description: "desc",
-          status: "DONE",
-          priority: "high",
-          owner: "ORCHESTRATOR",
-          depends_on: [],
-          tags: ["cli", "code", "git"],
-          verify: ["manual"],
-          verification: {
-            state: "ok",
-            updated_at: "2026-02-08T00:00:00.000Z",
-            updated_by: "TESTER",
-            note: "Verified: direct close should stage only the active task README.",
-          },
-          commit: { hash: implHash, message: "✨ R18Y1Q guard: impl" },
-          doc_version: 3,
-          doc_updated_at: "2026-02-08T00:00:00.000Z",
-          doc_updated_by: "TESTER",
-        },
-        "## Summary\n\nClose me\n",
-      );
-      const otherReadme = renderTaskReadme(
-        {
-          id: otherTaskId,
-          title: "Parallel active task",
-          result_summary: "",
-          description: "desc",
-          status: "DOING",
-          priority: "med",
-          owner: "CODER",
-          depends_on: [],
-          tags: ["code"],
-          verify: [],
-          verification: null,
-          commit: null,
-          doc_version: 3,
-          doc_updated_at: "2026-02-08T00:00:00.000Z",
-          doc_updated_by: "CODER",
-        },
-        "## Summary\n\nStill active\n",
-      );
-      await writeFile(path.join(closeTaskDir, "README.md"), closeReadme, "utf8");
-      await writeFile(path.join(otherTaskDir, "README.md"), otherReadme, "utf8");
-      await execFileAsync("git", ["add", ".agentplane/tasks"], { cwd: root });
-      await execFileAsync("git", ["commit", "-m", "task seed"], { cwd: root });
-
-      await writeFile(
-        path.join(closeTaskDir, "README.md"),
-        closeReadme.replace("## Summary\n\nClose me", "## Summary\n\nClose me now"),
-        "utf8",
-      );
-      await writeFile(
-        path.join(otherTaskDir, "README.md"),
-        otherReadme.replace("## Summary\n\nStill active", "## Summary\n\nStill active and dirty"),
-        "utf8",
-      );
-
-      const io = captureStdIO();
-      try {
-        const code = await runCli(["commit", closeTaskId, "--close", "--root", root]);
-        expect(code).toBe(0);
-        expect(io.stdout).toContain("committed");
-      } finally {
-        io.restore();
-      }
-
-      const showRes = await execFileAsync("git", ["show", "--name-only", "--format="], {
-        cwd: root,
-      });
-      const changed = showRes.stdout
-        .split("\n")
-        .map((s) => s.trim())
-        .filter(Boolean);
-      expect(changed).toEqual([`.agentplane/tasks/${closeTaskId}/README.md`]);
-
-      const statusRes = await execFileAsync("git", ["status", "--short", "--untracked-files=no"], {
-        cwd: root,
-      });
-      expect(statusRes.stdout).toContain(`.agentplane/tasks/${otherTaskId}/README.md`);
-      expect(statusRes.stdout).not.toContain(`.agentplane/tasks/${closeTaskId}/README.md`);
-    },
-    COMMIT_WRAPPER_CLOSE_CHECK_ONLY_TIMEOUT_MS,
-  );
-
-  it(
-    "commit wrapper --close supports --check-only",
-    async () => {
-      const root = await mkGitRepoRoot();
-      await writeDefaultConfig(root);
-      await configureGitUser(root);
-      const execFileAsync = promisify(execFile);
-
-      await writeFile(path.join(root, "seed.txt"), "seed\n", "utf8");
-      await execFileAsync("git", ["add", "seed.txt"], { cwd: root });
-      await execFileAsync("git", ["commit", "-m", "seed"], { cwd: root });
-      const implRes = await execFileAsync("git", ["rev-parse", "HEAD"], { cwd: root });
-      const implHash = implRes.stdout.trim();
-
-      const taskId = "202602081506-R18Y1Q";
-      const taskDir = path.join(root, ".agentplane", "tasks", taskId);
-      await mkdir(taskDir, { recursive: true });
-      const readme = renderTaskReadme(
-        {
-          id: taskId,
-          title: "Close preflight check",
-          result_summary: "check close",
-          description: "desc",
-          status: "DONE",
-          priority: "high",
-          owner: "ORCHESTRATOR",
-          depends_on: [],
-          tags: ["docs"],
-          verify: ["manual"],
-          verification: {
-            state: "ok",
-            updated_at: "2026-02-08T00:00:00.000Z",
-            updated_by: "TESTER",
-            note: "Verified: manual check",
-          },
-          commit: { hash: implHash, message: "✨ R18Y1Q docs: impl" },
-          doc_version: 3,
-          doc_updated_at: "2026-02-08T00:00:00.000Z",
-          doc_updated_by: "TESTER",
-        },
-        "## Summary\n\nTest task\n",
-      );
-      await writeFile(path.join(taskDir, "README.md"), readme, "utf8");
-
-      await writeFile(path.join(root, "noise.txt"), "noise\n", "utf8");
-      await execFileAsync("git", ["add", "noise.txt"], { cwd: root });
-
-      const beforeRes = await execFileAsync("git", ["rev-parse", "HEAD"], { cwd: root });
-      const before = beforeRes.stdout.trim();
-      const io = captureStdIO();
-      try {
-        const code = await runCli([
-          "commit",
-          taskId,
-          "--close",
-          "--check-only",
-          "--unstage-others",
-          "--root",
-          root,
-        ]);
-        expect(code).toBe(0);
-        expect(io.stdout).toContain("close preflight");
-        expect(io.stdout).toContain("would unstage 1 path(s)");
-      } finally {
-        io.restore();
-      }
-      const afterRes = await execFileAsync("git", ["rev-parse", "HEAD"], { cwd: root });
-      const after = afterRes.stdout.trim();
-      expect(after).toBe(before);
-    },
-    COMMIT_WRAPPER_CLOSE_CHECK_ONLY_TIMEOUT_MS,
-  );
-
-  it("commit wrapper requires a message", async () => {
-    const root = await mkGitRepoRoot();
-    const io = captureStdIO();
-    try {
-      const code = await runCli(["commit", "202601010101-ABCDEF", "--root", root]);
-      expect(code).toBe(2);
-      expect(io.stderr).toContain("Usage:");
-      expect(io.stderr).toContain("agentplane commit");
-    } finally {
-      io.restore();
-    }
-  });
-
-  it("commit wrapper rejects --unstage-others without --close", async () => {
-    const root = await mkGitRepoRoot();
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        "202601010101-ABCDEF",
-        "-m",
-        "✨ ABCDEF commit: docs update",
-        "--unstage-others",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(2);
-      expect(io.stderr).toContain("--unstage-others requires --close");
-    } finally {
-      io.restore();
-    }
-  });
-
-  it("commit wrapper rejects --check-only without --close", async () => {
-    const root = await mkGitRepoRoot();
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        "202601010101-ABCDEF",
-        "-m",
-        "✨ ABCDEF commit: docs update",
-        "--check-only",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(2);
-      expect(io.stderr).toContain("--check-only requires --close");
-    } finally {
-      io.restore();
-    }
-  });
-
-  it("commit wrapper requires a task id", async () => {
-    const root = await mkGitRepoRoot();
-    const io = captureStdIO();
-    try {
-      const code = await runCli(["commit", "--root", root]);
-      expect(code).toBe(2);
-      expect(io.stderr).toContain("Usage:");
-      expect(io.stderr).toContain("agentplane commit");
-    } finally {
-      io.restore();
-    }
-  });
-
-  it("commit wrapper rejects unknown flags", async () => {
-    const root = await mkGitRepoRoot();
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "commit",
-        "202601010101-ABCDEF",
-        "-m",
-        "✨ ABCDEF commit: run commit",
-        "--nope",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(2);
-      expect(io.stderr).toContain("Usage:");
-      expect(io.stderr).toContain("agentplane commit");
-    } finally {
-      io.restore();
-    }
-  });
-});
