+++ scripts/lib/test-inventory.mjs
@@ -92,14 +92,18 @@ export const SIGNIFICANT_COVERAGE_TARGETS = [
     tests: [
       "packages/agentplane/src/commands/guard/impl/commands.commit-close.unit.test.ts",
       "packages/agentplane/src/commands/guard/impl/commands.commit-non-close.unit.test.ts",
-      "packages/agentplane/src/cli/run-cli.core.guard.commit-wrapper.test.ts",
+      "packages/agentplane/src/cli/run-cli.core.guard.commit-wrapper.close.test.ts",
+      "packages/agentplane/src/cli/run-cli.core.guard.commit-wrapper.env.test.ts",
+      "packages/agentplane/src/cli/run-cli.core.guard.commit-wrapper.policy.test.ts",
+      "packages/agentplane/src/cli/run-cli.core.guard.commit-wrapper.refresh.test.ts",
     ],
   },
   {
     source: "packages/agentplane/src/commands/guard/impl/comment-commit.ts",
     tests: [
       "packages/agentplane/src/commands/guard/impl/comment-commit.test.ts",
-      "packages/agentplane/src/cli/run-cli.core.guard.commit-wrapper.test.ts",
+      "packages/agentplane/src/cli/run-cli.core.guard.commit-wrapper.close.test.ts",
+      "packages/agentplane/src/cli/run-cli.core.guard.commit-wrapper.env.test.ts",
     ],
   },
   {
